CodeSample of Web projects
AutoShop Ticket System: this is a PHP based web application that uses
MySql Database for managing an AutoShop.

Cryptogram: this is a simple game of deciphering a message which i did
for a real client (Thiagi.com) using JavaScript.

Chunks Puzzle: this is again a game which the chunks should be ordered
to form a meaningful phrase. in this we have implemented the Drag and
drop functionality using JavaScript and JQuery